refactor: market overview card
