+++ apps/main/src/llamalend/features/market-advanced-information/index.ts
@@ -3,5 +3,5 @@ export * from './MarketContractsSection'
 export * from './MarketLoanParameters'
 export * from './MarketParametersSection'
 export * from './AdvancedDetailsMetrics'
-export * from './MarketOverviewDetails'
+export * from './MarketOverviewCard'
 export * from './MarketAdvancedDetailsCard'
