+++ apps/main/src/loan/components/PageMintMarket/MintMarketPage.tsx
@@ -1,5 +1,6 @@
 import { useMemo, useState } from 'react'
 import { useConnection } from 'wagmi'
+import { MarketOverviewCard } from '@/llamalend/features/market-advanced-information/MarketOverviewCard'
 import { MarketContextProvider } from '@/llamalend/features/market-context'
 import { PositionDetailsComposite } from '@/llamalend/features/market-position-details'
 import { useIsInLiquidation } from '@/llamalend/features/market-position-details/hooks/useUserLiquidationStatus'
@@ -20,7 +21,7 @@ import type { Decimal } from '@primitives/decimal.utils'
 import { useCurve } from '@ui-kit/features/connect-wallet'
 import { useUserProfileStore } from '@ui-kit/features/user-profile'
 import { useParams } from '@ui-kit/hooks/router'
-import { useMarketMobileFormDrawer } from '@ui-kit/hooks/useFeatureFlags'
+import { useLlamaMarketDetailPageV2, useMarketMobileFormDrawer } from '@ui-kit/hooks/useFeatureFlags'
 import { t } from '@ui-kit/lib/i18n'
 import { ErrorPage } from '@ui-kit/pages/ErrorPage'
 import { MarketType, MarketRateType } from '@ui-kit/types/market'
@@ -36,6 +37,7 @@ export const MintMarketPage = () => {
   const { address } = useConnection()
   const [previewPrices, setPreviewPrices] = useState<Range<Decimal> | undefined>(undefined)
   const isMobileFormDrawer = useMarketMobileFormDrawer()
+  const isMarketDetailPageV2 = useLlamaMarketDetailPageV2()
 
   const marketQuery = useMintMarket({ chainId, rMarket: rCollateralId })
   const { data: market, isLoading: isMarketLoading, error: marketError } = marketQuery
@@ -103,6 +105,7 @@ export const MintMarketPage = () => {
       >
         <MarketBanners chainId={chainId} market={market} />
         <PositionDetailsComposite hasPosition={loanExists} events={collateralEvents} />
+        {isMarketDetailPageV2 && <MarketOverviewCard />}
         <MarketInformationComposite previewPrices={previewPrices} />
       </DetailPageLayout>
     </MarketContextProvider>
